+++ backend/canisters/community/c2c_client/src/lib.rs
@@ -16,6 +16,7 @@ generate_c2c_call!(summary_updates);
 // Updates
 generate_c2c_call!(c2c_bot_create_channel);
 generate_c2c_call!(c2c_bot_delete_channel);
+generate_c2c_call!(c2c_bot_delete_messages);
 generate_c2c_call!(c2c_bot_send_message);
 generate_c2c_call!(c2c_bot_subscribe_to_chat_events);
 generate_c2c_call!(c2c_create_proposals_channel);