+++ backend/canisters/local_user_index/api/src/updates/bot_delete_channel.rs
@@ -1,8 +1,7 @@
 use candid::CandidType;
-use oc_error_codes::OCError;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{AuthToken, ChannelId};
+use types::{AuthToken, ChannelId, UnitResult};
 
 #[ts_export(local_user_index, bot_delete_channel)]
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
@@ -11,15 +10,4 @@ pub struct Args {
     pub auth_token: AuthToken,
 }
 
-#[ts_export(local_user_index, bot_delete_channel)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-    ChannelNotFound,
-    FailedAuthentication(String),
-    InvalidRequest(String),
-    NotAuthorized,
-    Frozen,
-    C2CError(i32, String),
-    Error(OCError),
-}
+pub type Response = UnitResult;